+++ packages/twenty-front/src/modules/favorites/components/WorkspaceFavorites.tsx
@@ -2,17 +2,13 @@ import { useFilteredObjectMetadataItemsForWorkspaceFavorites } from '@/navigatio
 import { NavigationDrawerSectionForObjectMetadataItems } from '@/object-metadata/components/NavigationDrawerSectionForObjectMetadataItems';
 import { NavigationDrawerSectionForObjectMetadataItemsSkeletonLoader } from '@/object-metadata/components/NavigationDrawerSectionForObjectMetadataItemsSkeletonLoader';
 import { useIsPrefetchLoading } from '@/prefetch/hooks/useIsPrefetchLoading';
-import { usePrefetchedData } from '@/prefetch/hooks/usePrefetchedData';
-import { PrefetchKey } from '@/prefetch/types/PrefetchKey';
-import { View } from '@/views/types/View';
 
 export const WorkspaceFavorites = () => {
-  const { records: views } = usePrefetchedData<View>(PrefetchKey.AllViews);
-
   const { activeObjectMetadataItems: objectMetadataItemsToDisplay } =
     useFilteredObjectMetadataItemsForWorkspaceFavorites();
 
   const loading = useIsPrefetchLoading();
+
   if (loading) {
     return <NavigationDrawerSectionForObjectMetadataItemsSkeletonLoader />;
   }
@@ -21,7 +17,6 @@ export const WorkspaceFavorites = () => {
     <NavigationDrawerSectionForObjectMetadataItems
       sectionTitle={'Workspace'}
       objectMetadataItems={objectMetadataItemsToDisplay}
-      views={views}
       isRemote={false}
     />
   );
