+++ packages/twenty-front/src/modules/object-metadata/components/NavigationDrawerSectionForObjectMetadataItems.tsx
@@ -1,19 +1,13 @@
-import { useLastVisitedView } from '@/navigation/hooks/useLastVisitedView';
+import { NavigationDrawerItemForObjectMetadataItem } from '@/object-metadata/components/NavigationDrawerItemForObjectMetadataItem';
 import { ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
 import { NavigationDrawerAnimatedCollapseWrapper } from '@/ui/navigation/navigation-drawer/components/NavigationDrawerAnimatedCollapseWrapper';
-import { NavigationDrawerItem } from '@/ui/navigation/navigation-drawer/components/NavigationDrawerItem';
-import { NavigationDrawerItemsCollapsedContainer } from '@/ui/navigation/navigation-drawer/components/NavigationDrawerItemsCollapsedContainer';
 import { NavigationDrawerSection } from '@/ui/navigation/navigation-drawer/components/NavigationDrawerSection';
 import { NavigationDrawerSectionTitle } from '@/ui/navigation/navigation-drawer/components/NavigationDrawerSectionTitle';
-import { NavigationDrawerSubItem } from '@/ui/navigation/navigation-drawer/components/NavigationDrawerSubItem';
 import { useNavigationSection } from '@/ui/navigation/navigation-drawer/hooks/useNavigationSection';
-import { getNavigationSubItemState } from '@/ui/navigation/navigation-drawer/utils/getNavigationSubItemState';
-import { View } from '@/views/types/View';
-import { getObjectMetadataItemViews } from '@/views/utils/getObjectMetadataItemViews';
+import { ScrollWrapper } from '@/ui/utilities/scroll/components/ScrollWrapper';
 import styled from '@emotion/styled';
-import { useLocation } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
-import { useIcons } from 'twenty-ui';
+
 const ORDERED_STANDARD_OBJECTS = [
   'person',
   'company',
@@ -22,7 +16,7 @@ const ORDERED_STANDARD_OBJECTS = [
   'note',
 ];
 
-const ObjectsMetaDataItemsWrapper = styled.div`
+const StyledObjectsMetaDataItemsWrapper = styled.div`
   display: flex;
   flex-direction: column;
   gap: ${({ theme }) => theme.betweenSiblingsGap};
@@ -35,108 +29,46 @@ const ObjectsMetaDataItemsWrapper = styled.div`
 export const NavigationDrawerSectionForObjectMetadataItems = ({
   sectionTitle,
   isRemote,
-  views,
   objectMetadataItems,
 }: {
   sectionTitle: string;
   isRemote: boolean;
-  views: View[];
   objectMetadataItems: ObjectMetadataItem[];
 }) => {
   const { toggleNavigationSection, isNavigationSectionOpenState } =
     useNavigationSection('Objects' + (isRemote ? 'Remote' : 'Workspace'));
   const isNavigationSectionOpen = useRecoilValue(isNavigationSectionOpenState);
 
-  const { getIcon } = useIcons();
-  const currentPath = useLocation().pathname;
-  const { getLastVisitedViewIdFromObjectMetadataItemId } = useLastVisitedView();
-
-  const renderObjectMetadataItems = () => {
-    return [
-      ...objectMetadataItems
-        .filter((item) => ORDERED_STANDARD_OBJECTS.includes(item.nameSingular))
-        .sort((objectMetadataItemA, objectMetadataItemB) => {
-          const indexA = ORDERED_STANDARD_OBJECTS.indexOf(
-            objectMetadataItemA.nameSingular,
-          );
-          const indexB = ORDERED_STANDARD_OBJECTS.indexOf(
-            objectMetadataItemB.nameSingular,
-          );
-          if (indexA === -1 || indexB === -1) {
-            return objectMetadataItemA.nameSingular.localeCompare(
-              objectMetadataItemB.nameSingular,
-            );
-          }
-          return indexA - indexB;
-        }),
-      ...objectMetadataItems
-        .filter((item) => !ORDERED_STANDARD_OBJECTS.includes(item.nameSingular))
-        .sort((objectMetadataItemA, objectMetadataItemB) => {
-          return new Date(objectMetadataItemA.createdAt) <
-            new Date(objectMetadataItemB.createdAt)
-            ? 1
-            : -1;
-        }),
-    ].map((objectMetadataItem) => {
-      const objectMetadataViews = getObjectMetadataItemViews(
-        objectMetadataItem.id,
-        views,
+  const sortedStandardObjectMetadataItems = [...objectMetadataItems]
+    .filter((item) => ORDERED_STANDARD_OBJECTS.includes(item.nameSingular))
+    .sort((objectMetadataItemA, objectMetadataItemB) => {
+      const indexA = ORDERED_STANDARD_OBJECTS.indexOf(
+        objectMetadataItemA.nameSingular,
       );
-      const lastVisitedViewId = getLastVisitedViewIdFromObjectMetadataItemId(
-        objectMetadataItem.id,
+      const indexB = ORDERED_STANDARD_OBJECTS.indexOf(
+        objectMetadataItemB.nameSingular,
       );
-      const viewId = lastVisitedViewId ?? objectMetadataViews[0]?.id;
-
-      const navigationPath = `/objects/${objectMetadataItem.namePlural}${
-        viewId ? `?view=${viewId}` : ''
-      }`;
-
-      const isActive =
-        currentPath === `/objects/${objectMetadataItem.namePlural}`;
-      const shouldSubItemsBeDisplayed =
-        isActive && objectMetadataViews.length > 1;
-
-      const sortedObjectMetadataViews = [...objectMetadataViews].sort(
-        (viewA, viewB) =>
-          viewA.key === 'INDEX' ? -1 : viewA.position - viewB.position,
-      );
-
-      const selectedSubItemIndex = sortedObjectMetadataViews.findIndex(
-        (view) => viewId === view.id,
-      );
-
-      const subItemArrayLength = sortedObjectMetadataViews.length;
-
-      return (
-        <NavigationDrawerItemsCollapsedContainer
-          isGroup={shouldSubItemsBeDisplayed}
-        >
-          <NavigationDrawerItem
-            key={objectMetadataItem.id}
-            label={objectMetadataItem.labelPlural}
-            to={navigationPath}
-            Icon={getIcon(objectMetadataItem.icon)}
-            active={isActive}
-          />
-          {shouldSubItemsBeDisplayed &&
-            sortedObjectMetadataViews.map((view, index) => (
-              <NavigationDrawerSubItem
-                label={view.name}
-                to={`/objects/${objectMetadataItem.namePlural}?view=${view.id}`}
-                active={viewId === view.id}
-                subItemState={getNavigationSubItemState({
-                  index,
-                  arrayLength: subItemArrayLength,
-                  selectedIndex: selectedSubItemIndex,
-                })}
-                Icon={getIcon(view.icon)}
-                key={view.id}
-              />
-            ))}
-        </NavigationDrawerItemsCollapsedContainer>
-      );
+      if (indexA === -1 || indexB === -1) {
+        return objectMetadataItemA.nameSingular.localeCompare(
+          objectMetadataItemB.nameSingular,
+        );
+      }
+      return indexA - indexB;
     });
-  };
+
+  const sortedCustomObjectMetadataItems = [...objectMetadataItems]
+    .filter((item) => !ORDERED_STANDARD_OBJECTS.includes(item.nameSingular))
+    .sort((objectMetadataItemA, objectMetadataItemB) => {
+      return new Date(objectMetadataItemA.createdAt) <
+        new Date(objectMetadataItemB.createdAt)
+        ? 1
+        : -1;
+    });
+
+  const objectMetadataItemsForNavigationItems = [
+    ...sortedStandardObjectMetadataItems,
+    ...sortedCustomObjectMetadataItems,
+  ];
 
   return (
     objectMetadataItems.length > 0 && (
@@ -147,9 +79,18 @@ export const NavigationDrawerSectionForObjectMetadataItems = ({
             onClick={() => toggleNavigationSection()}
           />
         </NavigationDrawerAnimatedCollapseWrapper>
-        <ObjectsMetaDataItemsWrapper>
-          {isNavigationSectionOpen && renderObjectMetadataItems()}
-        </ObjectsMetaDataItemsWrapper>
+        <ScrollWrapper contextProviderName="navigationDrawer">
+          <StyledObjectsMetaDataItemsWrapper>
+            {isNavigationSectionOpen &&
+              objectMetadataItemsForNavigationItems.map(
+                (objectMetadataItem) => (
+                  <NavigationDrawerItemForObjectMetadataItem
+                    objectMetadataItem={objectMetadataItem}
+                  />
+                ),
+              )}
+          </StyledObjectsMetaDataItemsWrapper>
+        </ScrollWrapper>
       </NavigationDrawerSection>
     )
   );
