+++ packages/twenty-front/src/modules/ui/navigation/navigation-drawer/components/NavigationDrawerSection.tsx
@@ -7,7 +7,7 @@ const StyledSection = styled.div`
   width: 100%;
   margin-bottom: ${({ theme }) => theme.spacing(3)};
   flex-shrink: 1;
-  overflow: hidden
+  overflow: hidden;
 `;
 
 export { StyledSection as NavigationDrawerSection };
